refactor: put validation services to login service
put validation services to login service

MODULE-1-PART-3-gallery
