--- services/email-validation.service.js
@@ -1,16 +1,0 @@
-export class EmailValidationService {
-    validateEmail(email) {
-        let userEmail = {
-            isValid: false
-        }
-
-        let regexp = /\S+@\S+\.\S+/;
-        const isValid = regexp.test(email);
-
-        if (isValid) {
-            userEmail.isValid = true;
-        }
-
-        return userEmail;
-    }
-}
